권한에 따라 수정 삭제 버튼 표시 기능, 공지사항 삭제 시 확인창,
권한에 따라 수정 삭제 버튼 표시 기능,
공지사항 삭제 시 확인 창 띄워줌

+++ src/notice/components/NoticeDetailComponent.jsx
@@ -4,6 +4,8 @@ import { useState, useEffect } from "react";
 
 import { selectNoticeApi, deleteNoticeApi, BASE_URL } from "../api/NoticeApi";
 
+import { useUserInfo } from "../../store/useAuthStore";
+
 import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
 import "@/common/styles/DetailComponent.css";
@@ -19,10 +21,12 @@ export default function NoticeDetailComponent() {
 
     let navigate = useNavigate();
 
+    const user = useUserInfo();
+
     // 화면에 출력할 게시글 데이터를 담을 State 형 변수
     const [notice, setNotice] = useState({noticeId : "",
                                         title : "",
-                                        employee : {name : ""},
+                                        employee : { id: "", name : ""},
                                         content : "",
                                         createdAt : ""});
 
@@ -32,6 +36,14 @@ export default function NoticeDetailComponent() {
                                                    savedPath : "",
                                                    extension: "",
                                                    notices : {noticeId : ""}});
+
+    const userId = user.id || "";
+    const canUpdateDeleteNotice = Boolean(
+        userId && 
+        notice.employee.id &&
+        // '같은 아이디의 작성자'임에도 수정/삭제 버튼이 안 뜨는 버그를 예방 Number 로 타입 통일
+        Number(userId)=== Number(notice.employee.id)
+    );
                                                 
     // 이 컴포넌트가 브라우저에 다 로딩된 후 딱 한번 실행할 구문
     useEffect(() => {
@@ -74,6 +86,13 @@ export default function NoticeDetailComponent() {
     // 삭제하기 버튼 클릭 시 실행할 이벤트 핸들러 함수
     const deleteNotice = async () => {
 
+        // 1. 사용자 확인 창 띄우기
+        const isConfirmed = window.confirm("정말 삭제하시겠습니까?");
+    
+        // 취소를 누른 경우 함수 종료 (삭제 안 함)
+        if (!isConfirmed) return;
+        
+        // 2. 확인 눌렀을때
         try {
 
             const response = await deleteNoticeApi(noticeId);
@@ -99,13 +118,13 @@ export default function NoticeDetailComponent() {
     };
 
     return (
-        <div className="detail-page">
+        <div className="detail-page w-full">
             <PageHeader 
-            title="공지사항 상세" 
+            title="공지사항 상세조회" 
             description="공지사항 정보를 상세조회 합니다." 
             icon={Megaphone}/>
         
-            <div className="detail-page">
+            <div className="detail-page w-full">
                 <section data-detail-section="info">
                     <Card className="detail-info-card">
                         <CardHeader>
@@ -125,12 +144,12 @@ export default function NoticeDetailComponent() {
                                     <dd>{ notice.employee.name }</dd>
                                 </div>
 
-                                <div>
+                                <div className="full-row">
                                     <dt width="130">작성일</dt>
                                     <dd>{ notice.createdAt?.substring(0, 10) }</dd>
                                 </div>
-                                <div></div>
-                                <div>
+                                
+                                <div className="full-row">
                                     <dt>내용</dt>
                                     <dd>
                                         <p style={ {height : "300px"} }>
@@ -138,8 +157,8 @@ export default function NoticeDetailComponent() {
                                         </p>
                                     </dd>
                                 </div>
-                                <div></div>
-                                <div>
+                                
+                                <div className="full-row">
                                     <dt>첨부파일</dt>
                                     <dd>
                                         {
@@ -154,13 +173,14 @@ export default function NoticeDetailComponent() {
                                 </div>
                             </dl>
                         </CardContent>
-
+                     
                         <CardFooter>
                             <Button size="lg" className="cursor-pointer"
                                     onClick={ () => { navigate("/notices/list"); } }>
                                 목록으로
                             </Button>
-                            
+                        {canUpdateDeleteNotice && (
+                           <>
                             <Button size="lg" className="cursor-pointer"
                                     onClick={ () => { navigate(`/notices/${noticeId}/edit`, {state : {noticeId}}); } }>
                                 편집
@@ -170,7 +190,10 @@ export default function NoticeDetailComponent() {
                                     onClick={ deleteNotice }>
                                 삭제
                             </Button>
+                           </> 
+                        )}
                         </CardFooter>
+                        
                     </Card>
                 </section>
             </div>
